Expose heapsort algorithms in a shared header
Fix #22011
